+++ spec/rubocop/cli_spec.rb
@@ -1443,10 +1443,20 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
       YAML
 
       expect(cli.run(%w[--format simple example])).to eq(1)
-      expect($stderr.string)
-        .to eq(['Warning: unrecognized parameter Metrics/LineLength:Min ' \
-                'found in example/.rubocop.yml',
-                ''].join("\n"))
+
+      expect($stderr.string).to eq(<<-RESULT.strip_margin('|'))
+        |Warning: Metrics/LineLength does not support Min parameter.
+        |
+        |Supported parameters are:
+        |
+        |  - Enabled
+        |  - Max
+        |  - AllowHeredoc
+        |  - AllowURI
+        |  - URISchemes
+        |  - IgnoreCopDirectives
+        |  - IgnoredPatterns
+      RESULT
     end
 
     it 'prints an error message for an unrecognized EnforcedStyle' do